html/browsers/browsing-the-web/navigating-across-documents/012-timeout.html

<!doctype html>
<title>Link with onclickwith delayed (timeout) document.write and href navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<a target="test" onclick="setTimeout(() => document.write('<script>parent.postMessage(&quot;write&quot;, &quot;*&quot;)</script>'), 2000);" href="href.html">Test</a>
<script>
/*---
specid: navigating-across-documents
steps: [6]
info: |
  Cancel any preexisting but not yet mature attempt to navigate browsingContext, including canceling any instances of the fetch algorithm started by those attempts. If one of those attempts has already created and initialized a new Document object, abort that Document also. (Navigation attempts that have matured already have session history entries, and are therefore handled during the update the session history with the new page algorithm, later.)

  Step 12 contains https://html.spec.whatwg.org/multipage/browsing-the-web.html#javascript-protocol

  - https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#concept-form-submit
  - https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#submit-get-action
  - https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#plan-to-navigate
  - https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#planned-navigation

description: |
  1. This test creates an anchor with both an onclick and href.
  2. The test "clicks" the anchor, triggering the onclick handler operation, which will attempt to document.write (after a 2s delay), followed by navigating the iframe to the anchor's `href=""`.

---*/

var t = async_test(undefined, {
  timeout: 4000
});

t.step(function() {
  document.getElementsByTagName("a")[0].click();
});

onmessage = t.step_func(function(e) {
  assert_equals(e.data, "href");
  t.done();
});

</script>
